Refactoring code
Updated XML namespaces and Main class
