+++ src/main/resources/mappers/customerMapper.xml
@@ -3,7 +3,7 @@
         PUBLIC "-//mybatis.ord//DTD Mapper 3.0//EN"
         "http://mybatis.org//dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="myBatis.CustomerMapper">
+<mapper namespace="com.solvd.delivery.service.ICustomerService.CustomerMapper">
 
     <select id="getByID" parameterType="int" resultMap="customerResultMap">
         SELECT * FROM Customers WHERE ID = #{id};
